LayoutTests/imported/w3c/web-platform-tests/requestidlecallback/deadline-max-timeout-dynamic-expected.txt

Test of requestIdleCallback deadline behavior

The test can pass accidentally as idle deadlines have a maximum but they can always be shorter. It runs multiple times to expose potential failures.


FAIL Check that the deadline is changed if there is a new timeout from within requestIdleCallback. assert_less_than_equal: expected a number less than or equal to 10 but got 95
